requirements.txt
# Databricks Apps: pip install -r requirements.txt
# Same pins as the Databricks Gradio template (gradio + huggingface_hub + pandas).
# cache-bust: 2026-04-05 — Nyaya Sahayak feature expansion
gradio~=4.44.0
gradio-client==1.3.0
huggingface-hub~=0.35.3
pandas~=2.2.3
databricks-sdk>=0.20
# MLflow: graceful degradation — app works without it, but logs queries when available.
mlflow>=2.12.0
-e ".[rag,rag_embed,app]"
